Update getNCI60.R
Not having the `sort=FALSE` means that the result is sorted by the by.x column. This leads to a mismatch of the rows in the merge data table from the original rowData, so all the added gene symbols, ENSG ids, etc are out of order in the current PSet.